Secured login/admin only views

+++ products/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
@@ -66,8 +67,13 @@ def product_detail(request, product_id):  # view to show individual products
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     # add product to store
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only store users can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -87,8 +93,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     # Edit a product
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only store users can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -111,7 +122,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    # Delete a product
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only store users can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'Product {product.name} Successfully deleted')
